refactor: refactor sign up component , add useAuth.jsx file to Hooks folder Relates #34
