+++ client/src/Components/ProtectedRoute .jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Navigate } from 'react-router-dom';
-import { useAuth } from '../useAuth';
+import { useAuth } from '../Hooks/useAuth';
 
 function ProtectedRoute({ redirectPath = '/', children }) {
   const { user, loading } = useAuth();
